requirement.txt
python -m pip install --upgrade pip
python -m pip install --upgrade pytest
python -m pip install --upgrade pluggy
pytest-html
pandas openpyxl
